Fixed redirect without reload not triggering UI display
Because in many cases we end up at the Following page via template rendering and not page loading the "matches" attribute in manifest.json is insufficient to display the filter UI in every case where we arrive at the Following page.

I added an interval that checks every .5 seconds if we're on the following page (I would imagine this is super performance unintensive).  If we're not it continues.  If we are it sets up the UI and clears the interval to stop running.